use Nix Flakes over Niv
Pa'i and everything associated can be compiled with:

	nix build --impure

shell.nix has been replace with Nix's inferences from the default
package output. It can be accessed with:

	nix shell --impure

docker.nix has been replaced with a *derivation* that wraps
buildLayeredImage. It installs the image to $out/docker/, and can
be built with:

	nix build .#docker --impure

All of the above commands assume that the installed Nix version
supports flakes. The `impure` option is necessary as not everything
down the dependency list is built with flakes, and there are some
impure/non-hermetic calls to `builtins.currentSystem`.

Note that this currently only builds on x86_64-linux hosts. I don't
have any other kernels or architectures to test with on-hand, so I
can't verify build correctness for anything else --- this is
*technically* the proper behavior for Flakes.